Remove op-labels feature
Closes #1775
